Address Phase 1 review: exit code type, CLI smoke test, venv docs
- Change Report::exit_code() to return u8 so the CLI no longer needs a
  fallible conversion (was silently mapping out-of-range to exit 2).
- Add end-to-end smoke test for the zorilla binary covering the empty
  and populated-tree Phase 1 summary lines.
- Document that maturin develop requires an activated venv (or exported
  VIRTUAL_ENV) in CLAUDE.md and README.md — flagged during the check step.

+++ crates/zorilla-core/src/report.rs
@@ -42,9 +42,10 @@ impl Report {
         format!("{} findings in {} files discovered.", self.findings.len(), self.files_discovered)
     }
 
-    /// Process exit code: `0` clean, `1` findings, `2` error (set by caller).
-    pub fn exit_code(&self) -> i32 {
-        i32::from(!self.findings.is_empty())
+    /// Process exit code: `0` clean, `1` findings. `2` (error) is set by
+    /// the CLI when it can't produce a `Report` at all.
+    pub fn exit_code(&self) -> u8 {
+        u8::from(!self.findings.is_empty())
     }
 }
